refactor(gyro_odometer): move transform_covariance to its only caller
The node applies it to the IMU sample it has just transformed, and
nothing in the logic class calls it. It moves next to transform_imu(),
and its test moves with it. What the logic class expects of an IMU
sample is now stated on input_imu().
