+++ localization/autoware_gyro_odometer/test/test_gyro_odometer_node_characterization.cpp
@@ -27,6 +27,7 @@
 
 #include <gtest/gtest.h>
 
+#include <array>
 #include <chrono>
 #include <functional>
 #include <map>
@@ -394,6 +395,33 @@ TEST_F(GyroOdometerNodeCharacterization, AccumulatedVehicleTwistsAreAveragedInto
   EXPECT_EQ(reported_int(diagnostics, "vehicle_twist_queue_size"), 2);
   EXPECT_EQ(reported_int(diagnostics, "imu_queue_size"), 1);
 }
+
+// transform_covariance: the maximum diagonal term is written to every diagonal term, off-diagonals
+// are zeroed.
+TEST(GyroOdometer, TransformCovariancePicksMaxDiagonalAndZerosOffDiagonals)
+{
+  std::array<double, 9> cov = {};
+  cov[COV_IDX_XYZ::X_X] = 1.0;
+  cov[COV_IDX_XYZ::Y_Y] = 5.0;  // max
+  cov[COV_IDX_XYZ::Z_Z] = 3.0;
+  // pollute off-diagonals to make sure they are dropped
+  cov[COV_IDX_XYZ::X_Y] = 42.0;
+  cov[COV_IDX_XYZ::Z_X] = -7.0;
+
+  const std::array<double, 9> out = transform_covariance(cov);
+
+  EXPECT_DOUBLE_EQ(out[COV_IDX_XYZ::X_X], 5.0);
+  EXPECT_DOUBLE_EQ(out[COV_IDX_XYZ::Y_Y], 5.0);
+  EXPECT_DOUBLE_EQ(out[COV_IDX_XYZ::Z_Z], 5.0);
+  EXPECT_DOUBLE_EQ(out[COV_IDX_XYZ::X_Y], 0.0);
+  EXPECT_DOUBLE_EQ(out[COV_IDX_XYZ::X_Z], 0.0);
+  EXPECT_DOUBLE_EQ(out[COV_IDX_XYZ::Y_X], 0.0);
+  EXPECT_DOUBLE_EQ(out[COV_IDX_XYZ::Y_Z], 0.0);
+  EXPECT_DOUBLE_EQ(out[COV_IDX_XYZ::Z_X], 0.0);
+  EXPECT_DOUBLE_EQ(out[COV_IDX_XYZ::Z_Y], 0.0);
+}
+
+// fuse_twist: means over multiple entries, covariance reduction by queue size, fixed Y_Y/Z_Z, and
 
 // A completed pair puts the longitudinal velocity of the vehicle twist and the angular velocity of
 // the IMU on all four output topics.
